refactor: split agent task matrix module
